move class definitions for Control, Model and SolveHandle to separate file
